+++ src/test/java/com/picimako/mockitools/intention/ConvertSpyCallToFieldIntentionTest.java
@@ -159,6 +159,34 @@ class ConvertSpyCallToFieldIntentionTest extends MockitoolsIntentionTestBase {
                 }""");
     }
 
+    @Test
+    void testNotAvailableForGenericInferredMockitoSpyWithVarKeyword() {
+        checkIntentionIsNotAvailable(
+            """
+                import org.mockito.Mockito;
+
+                public class NotAvailable {
+                    public void testMethod() {
+                        var spy = Mockito.sp<caret>y();
+                    }
+                }""");
+    }
+
+    @Test
+    void testAvailableForGenericInferredMockitoSpyWithExplicitType() {
+        checkIntentionIsAvailable(
+            """
+                import org.mockito.Mockito;
+
+                public class Available {
+                    public void testMethod() {
+                        Mockable spy = Mockito.sp<caret>y();
+                    }
+                
+                    public static final class Mockable { }
+                }""");
+    }
+
     //Conversion - standalone spy call + new expression
 
     @Test
@@ -950,6 +978,36 @@ class ConvertSpyCallToFieldIntentionTest extends MockitoolsIntentionTestBase {
                 public class ConversionTest {
                     public void testMethod() {
                         var spy = Mockito.s<caret>py(Mockable.class);
+                    }
+
+                    public static final class Mockable { }
+                }""",
+            """
+                import org.mockito.Mockito;
+                import org.mockito.Spy;
+
+                public class ConversionTest {
+                    @Spy
+                    Mockable spy;
+
+                    public void testMethod() {
+                    }
+
+                    public static final class Mockable { }
+                }""");
+    }
+
+    //Conversion - generic inferred spy creation
+
+    @Test
+    void testConvertsGenericInferredMockitoSpy() {
+        checkIntentionRun(
+            """
+                import org.mockito.Mockito;
+
+                public class ConversionTest {
+                    public void testMethod() {
+                        Mockable spy = Mockito.s<caret>py();
                     }
 
                     public static final class Mockable { }